+++ BoxGame/Timed_Text.h
@@ -5,7 +5,7 @@
 class Timed_Text : public GUI_Object
 {
 public:
-	Timed_Text(Game& game, std::string txt, sf::Color color, sf::Vector2f position, size_t size, size_t lifelength);
+	Timed_Text(Game& game, std::string txt, sf::Color color, sf::Vector2f position, size_t size, float lifelength);
 
 	void draw() override;
 	bool update() override;
@@ -13,5 +13,5 @@ public:
 private:
 	sf::Text text;
 	sf::Clock clock;
-	size_t lifelength;
+	float lifelength;
 };
